Move type symbol impls to symbols package
