+++ compiler/ballerina-lang/spotbugs-exclude.xml
@@ -88,10 +88,6 @@
     </Match>
     <Match>
         <Package name="io.ballerina.compiler.api.impl.symbols" />
-        <Bug pattern="BC_UNCONFIRMED_CAST" />
-    </Match>
-    <Match>
-        <Package name="io.ballerina.compiler.api.impl.types" />
         <Bug pattern="BC_UNCONFIRMED_CAST, BC_UNCONFIRMED_CAST_OF_RETURN_VALUE" />
     </Match>
     <Match>
