+++ compiler/ballerina-lang/src/main/java/io/ballerina/compiler/api/impl/TypesFactory.java
@@ -18,21 +18,21 @@
 package io.ballerina.compiler.api.impl;
 
 import io.ballerina.compiler.api.ModuleID;
-import io.ballerina.compiler.api.impl.types.BallerinaArrayTypeSymbol;
-import io.ballerina.compiler.api.impl.types.BallerinaErrorTypeSymbol;
-import io.ballerina.compiler.api.impl.types.BallerinaFunctionTypeSymbol;
-import io.ballerina.compiler.api.impl.types.BallerinaFutureTypeSymbol;
-import io.ballerina.compiler.api.impl.types.BallerinaMapTypeSymbol;
-import io.ballerina.compiler.api.impl.types.BallerinaNilTypeSymbol;
-import io.ballerina.compiler.api.impl.types.BallerinaObjectTypeSymbol;
-import io.ballerina.compiler.api.impl.types.BallerinaRecordTypeSymbol;
-import io.ballerina.compiler.api.impl.types.BallerinaSimpleTypeSymbol;
-import io.ballerina.compiler.api.impl.types.BallerinaSingletonTypeSymbol;
-import io.ballerina.compiler.api.impl.types.BallerinaStreamTypeSymbol;
-import io.ballerina.compiler.api.impl.types.BallerinaTupleTypeSymbol;
-import io.ballerina.compiler.api.impl.types.BallerinaTypeDescTypeSymbol;
-import io.ballerina.compiler.api.impl.types.BallerinaTypeReferenceTypeSymbol;
-import io.ballerina.compiler.api.impl.types.BallerinaUnionTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaArrayTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaErrorTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaFunctionTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaFutureTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaMapTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaNilTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaObjectTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaRecordTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaSimpleTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaSingletonTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaStreamTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaTupleTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaTypeDescTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaTypeReferenceTypeSymbol;
+import io.ballerina.compiler.api.impl.symbols.BallerinaUnionTypeSymbol;
 import io.ballerina.compiler.api.symbols.TypeDescKind;
 import io.ballerina.compiler.api.symbols.TypeSymbol;
 import org.ballerinalang.model.types.TypeKind;
