+++ compiler/ballerina-lang/src/main/java/module-info.java
@@ -12,7 +12,6 @@ module io.ballerina.lang {
     requires io.ballerina.tools.api;
     exports io.ballerina.compiler.api;
     exports io.ballerina.compiler.api.symbols;
-    exports io.ballerina.compiler.api.types;
     exports io.ballerina.compiler.api.impl;
     exports org.wso2.ballerinalang.compiler.util;
     exports org.ballerinalang.toml.model;
